DRY: Consolidate glyph status persistence into generic storage helpers
Eliminates duplication of save/load logic across IX and Prompt glyphs
by introducing a generic GlyphStatus type and reusable storage functions.

Changes:
- Created web/ts/components/glyph/glyph-storage.ts with:
  - Base GlyphStatus interface (state, message, timestamp)
  - ExecutableGlyphStatus extension (scheduledJobId, executionId)
  - Generic saveGlyphStatus<T> and loadGlyphStatus<T> functions
- Refactored ix-glyph.ts to use ExecutableGlyphStatus
- Refactored prompt-glyph.ts to use base GlyphStatus
- Removed 40+ lines of duplicate save/load logic per glyph type

Preserves exact behavior while providing a foundation for future glyph
types to reuse status persistence without code duplication.
